+++ Trident.Core/Memory/GamePak/GamePak.cs
@@ -67,7 +67,7 @@ namespace Trident.Core.Memory.GamePak
             if (TAccess.IsLower)
                 if (IsGPIOAddress(address) && _gpio.Readable) return _gpio.Read(address);
             else
-                // EEPROM does not use the address parameter; it is a purely serial device. Pass in uint.MaxValue to signify that it doesn't matter.
+                // EEPROM does not use the address parameter; it is a purely serial device. Pass in 0xFFFFFFFF to signify that it doesn't matter.
                 if (IsEEPROMAddress(address)) return _backupDevice.Read(0xFFFF_FFFF);
 
             // Seqential accesses from one address will not overwrite the latch! We should instead use the auto-incremented value.
@@ -119,7 +119,7 @@ namespace Trident.Core.Memory.GamePak
             {
                 ushort low = (ushort)(_romAddress >> 1);
                 ushort top = (ushort)(low + 1);
-                value = (ushort)((top << 16) | low);
+                value = (uint)((top << 16) | low);
             }
 
             // The GBA ROM uses an internal latch which automatically increments.
